+++ src/main/java/ru/javarush/country/Main.java
@@ -1,4 +1,4 @@
-package ru.javarush;
+package ru.javarush.country;
 
 public class Main {
     public static void main(String[] args) {
