+++ src/main/java/ru/javarush/country/dao/GenericDao.java
@@ -1,0 +1,5 @@
+package ru.javarush.country.dao;
+
+public abstract class GenericDao {
+
+}
