--- src/test/java/ru/javarush/MainTest.java
@@ -1,5 +1,0 @@
-package ru.javarush;
-
-public class MainTest {
-
-}
