+++ src/test/java/ru/javarush/country/MainTest.java
@@ -1,0 +1,5 @@
+package ru.javarush.country;
+
+public class MainTest {
+
+}
